Refactor components and styles
