+++ src/components/Overlay/Navbar/style.css
@@ -1,0 +1,70 @@
+.container {
+  max-width: 1140px;
+  width: 100%;
+  margin: 0 auto;
+  padding: 0 32px;
+}
+
+header {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  margin: 32px 0;
+}
+
+nav > ul {
+  list-style: none;
+  display: flex;
+  flex-direction: row;
+  column-gap: 32px;
+}
+
+.navigation {
+  display: flex;
+  flex-direction: row;
+  justify-content: end;
+  column-gap: 80px;
+}
+
+.navigation > ul {
+  flex-direction: row;
+  column-gap: 32px;
+}
+
+button {
+  font-family: "Inter", sans-serif;
+  letter-spacing: 1.6px;
+  font-size: 18px;
+  color: rgb(164, 164, 164);
+  background: none;
+  border: none;
+  cursor: pointer;
+}
+
+button > svg {
+  display: block;
+  height: 24px;
+  width: 24px;
+  stroke: rgb(164, 164, 164);
+  cursor: pointer;
+}
+
+.hamburger {
+  display: none;
+}
+
+.hamburger > svg {
+  fill: rgb(164, 164, 164);
+}
+
+/* For screens less than 700px */
+@media only screen and (max-width: 700px) {
+  .navigation > ul {
+    display: none;
+  }
+
+  .hamburger {
+    display: block;
+  }
+}
