+++ src/style.css
@@ -156,185 +156,3 @@ a:hover {
   border: 1px solid #000;
   padding: 6px 12px;
 }
-
-/* Navigation */
-
-.container {
-  max-width: 1140px;
-  width: 100%;
-  margin: 0 auto;
-  padding: 0 32px;
-}
-
-header {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  margin: 32px 0;
-}
-
-nav > ul {
-  list-style: none;
-  display: flex;
-  flex-direction: row;
-  column-gap: 32px;
-}
-
-.navigation {
-  display: flex;
-  flex-direction: row;
-  justify-content: end;
-  column-gap: 80px;
-}
-
-.navigation > ul {
-  flex-direction: row;
-  column-gap: 32px;
-}
-
-button {
-  font-family: "Inter", sans-serif;
-  letter-spacing: 1.6px;
-  font-size: 18px;
-  color: rgb(164, 164, 164);
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-
-button > svg {
-  display: block;
-  height: 24px;
-  width: 24px;
-  stroke: rgb(164, 164, 164);
-  cursor: pointer;
-}
-
-.hamburger {
-  display: none;
-}
-
-.hamburger > svg {
-  fill: rgb(164, 164, 164);
-}
-
-.mobileNavOverlay {
-  display: none;
-  position: absolute;
-  top: 100%;
-  left: 2.5%;
-  width: 95%;
-  margin-top: 32px;
-  border-radius: 20px;
-  background: rgb(122, 160, 216);
-  background: linear-gradient(
-    133deg,
-    rgba(122, 160, 216, 0.29) 0%,
-    rgba(13, 34, 67, 0.39) 100%
-  );
-  box-shadow: 4px 4px 6px rgba(0, 0, 0, 0.125);
-}
-
-.mobileNav {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 48px 32px;
-  row-gap: 48px;
-}
-
-.mobileNav > ul {
-  flex-direction: column;
-  align-items: center;
-  row-gap: 24px;
-}
-
-.mobileNav > ul:nth-of-type(2) {
-  flex-direction: row;
-}
-
-.btnContainer {
-  display: flex;
-  flex-direction: column;
-  row-gap: 24px;
-  align-items: start;
-}
-
-.sectionBtn {
-  display: flex;
-  column-gap: 18px;
-  align-items: center;
-  font-family: "Inter", sans-serif;
-  font-size: 24px;
-  font-weight: 300;
-  text-transform: uppercase;
-  color: rgb(164, 164, 164);
-}
-
-/* Section details */
-
-.sectionOverlay {
-  position: absolute;
-  top: 50%;
-  left: 32px;
-  transform: translateY(-50%);
-  width: 600px;
-  border-radius: 20px;
-  padding: 32px 48px;
-  display: flex;
-  flex-direction: column;
-  row-gap: 100px;
-  background: rgb(122, 160, 216);
-  background: linear-gradient(
-    133deg,
-    rgba(122, 160, 216, 0.29) 0%,
-    rgba(13, 34, 67, 0.39) 100%
-  );
-  box-shadow: 4px 4px 6px rgba(0, 0, 0, 0.125);
-}
-
-.sectionText > h1 {
-  font-size: 48px;
-  font-weight: 300;
-  margin-bottom: 20px;
-}
-
-.sectionText > p {
-  font-size: 20px;
-  font-weight: 300;
-  line-height: 30px;
-}
-
-.temp-btn {
-  position: absolute;
-  top: 16px;
-  left: 100%;
-  margin-left: 8px;
-  padding: 16px;
-  border-radius: 20px;
-  background: linear-gradient(
-    133deg,
-    rgba(122, 160, 216, 0.29) 0%,
-    rgba(13, 34, 67, 0.39) 100%
-  );
-}
-
-/* For screens less than 700px */
-@media only screen and (max-width: 700px) {
-  .navigation > ul {
-    display: none;
-  }
-
-  .hamburger {
-    display: block;
-  }
-
-  .mobileNavOverlay {
-    display: block;
-  }
-
-  button {
-    text-transform: uppercase;
-  }
-}
